+++ src/org/netbeans/modules/php/blade/editor/ResourceUtilities.java
@@ -39,7 +39,6 @@ public final class ResourceUtilities {
     public static final String CUSTOM_HTML_ICON = "org/netbeans/modules/html/custom/resources/custom_html_element.png"; //NOI18N
     public static final String XML_ATTRIBUTE_ICON = "org/netbeans/modules/xml/schema/completion/resources/attribute.png"; //NOI18N
     
-    
     private ResourceUtilities() {
         
     }