+++ src/org/netbeans/modules/php/blade/editor/lexer/BladeTokenId.java
@@ -45,7 +45,7 @@ public enum BladeTokenId implements TokenId {
     BLADE_PAREN("token"), // NOI18N
     BLADE_COMPONENT_ATTRIBUTE("blade_comp_attribute"), // NOI18N
     HTML("html"), // NOI18N
-    WS_D("html"), // NOI18N
+    WS_D("whitespace"), // NOI18N
     BLADE_DIRECTIVE_UNKNOWN("at_string"), // NOI18N
     PHP_BLADE_EXPRESSION("blade_php"), // NOI18N
     PHP_BLADE_ECHO_EXPR("blade_php"), // NOI18N