+++ .github/workflows/CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## What's changed
 
-- project indexing defensive check for project load
-- pom.xml deployment updates
+- Relative declaration finder for blade files of components inside `view/components` or `view/livewire`
+- fix project blade config check on stand alone php file (null project) on hyperlink action
+- include `@overwrite` brace matcher back search to `@section`
+- fix css unit '%' style attribute error near blade tag