+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [3.0.0.39]
+
+- Relative declaration finder for blade files of components inside `view/components` or `view/livewire`
+- fix project blade config check on stand alone php file (null project) on hyperlink action
+- include `@overwrite` brace matcher back search to `@section`
+- fix css unit '%' style attribute error near blade tag
+
 ## [3.0.0.38]
 - project indexing defensive check for project load
 - pom.xml deployment updates